code cleanup: remove redundant casts

+++ source/gameengine/BlenderRoutines/KX_BlenderCanvas.cpp
@@ -119,14 +119,14 @@ int KX_BlenderCanvas::GetHeight(
 
 int KX_BlenderCanvas::GetMouseX(int x)
 {
-	float left = GetWindowArea().GetLeft();
-	return float(x - (left - m_area_left));
+	int left = GetWindowArea().GetLeft();
+	return x - (left - m_area_left);
 }
 
 int KX_BlenderCanvas::GetMouseY(int y)
 {
-	float top = GetWindowArea().GetTop();
-	return float(y - (m_area_top - top));
+	int top = GetWindowArea().GetTop();
+	return y - (m_area_top - top);
 }
 
 float KX_BlenderCanvas::GetMouseNormalizedX(int x)
